+++ packages/blade/src/components/Badge/StyledBadge.web.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { getStyledBadgeStyles } from './getStyledBadgeStyles';
-import type { StyledBadgeProps } from './StyledBadge.d';
+import type { StyledBadgeProps } from './types';
 
 const StyledBadge = styled.div<StyledBadgeProps>((props) => ({
   ...getStyledBadgeStyles(props),