+++ packages/blade/src/components/Form/Selector/types.ts
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export type SelectorLabelProps = {
+  children: React.ReactNode;
+  /**
+   * Pass only on react-native
+   */
+  inputProps: any; // Partial<Extract<InputProps, { onPress: any }>>;
+};