+++ packages/blade/src/components/Icons/_Svg/Circle/types.ts
@@ -1,0 +1,9 @@
+export type CircleProps = {
+  fill?: string;
+  stroke?: string;
+  r: string;
+  cx: string;
+  cy: string;
+  x?: string;
+  y?: string;
+};