+++ packages/blade/src/components/Icons/_Svg/Rect/Rect.native.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 import { Rect as RectNative } from 'react-native-svg';
-import type { RectProps } from './Rect';
+import type { RectProps } from './types';
 
 const Rect = ({ height, width, rx, ry, x, y }: RectProps): ReactElement => {
   return <RectNative height={height} width={width} rx={rx} ry={ry} x={x} y={y} />;